Add files via upload
This is the result of a coding challenge that I set my most talented Computing student: use Minecraft to physically model the electric field potential of a gold nucleus and a nearby carbon nucleus. This was my first FORTRAN coding assignment as a first year Physics undergraduate at university, and my Year 11 student cracked it in an afternoon. #BornToCode

You can customise this script by using a Periodic table to find out the electric charge values of different elemental nuclei, and the whole thing will scale accordingly. You can also try adding in extra nuclei to see the effect on the overall shape of the electric field.

Enjoy!